server/requirements.txt
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
channels==4.3.2
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Django==5.2.9
django-channels==0.7.0
django-cors-headers==4.9.0
django-restframework==0.0.1
djangorestframework==3.16.1
h11==0.16.0
httptools==0.7.1
idna==3.11
oauthlib==3.3.1
pillow==12.0.0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
